pyproject.toml
[tool.poetry]
name = "dharma-chatbot"
version = "0.1.0"
description = ""
authors = ["gsajko <grzegorz.sajko@protonmail.com>"]
readme = "README.md"
packages = [{include = "dharma_chatbot"}]

[tool.poetry.dependencies]
python = "^3.10"
pyvespa = "^0.39.0"
langchain = "^0.0.352"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
ruff = "^0.1.8"
ipykernel = "^6.27.1"



[tool.black]
line-length = 88
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.nox
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | data/fuel_data
  )/
)
'''
[tool.ruff]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


